!working: Cheek crud implemented
Due to major changes to server accuring during implementation of client, some separation needed to be made to ensure separation of branches.

This commit only contains SOME of the relevant parts of the new implementation, meaning that this commit still MISSES some of the changes needed for the application to run. Even though the necessary logic for the cheek crud exists.

After applying and commiting other changes to their respective repositories, there will be a new commit on this branch ensuring full compatibility.